+++ DatadogTrace/Tests/TracingURLSessionHandlerTests.swift
@@ -244,9 +244,11 @@ class TracingURLSessionHandlerTests: XCTestCase {
         let isKept: Bool = true
 
         // Given
+        // TODO: RUM-13455 - Trace requires metrics mode for now
         let interception = URLSessionTaskInterception(
             request: .mockAny(),
-            isFirstParty: true
+            isFirstParty: true,
+            trackingMode: .metrics
         )
         interception.register(response: .mockAny(), error: nil)
         interception.register(
@@ -290,7 +292,8 @@ class TracingURLSessionHandlerTests: XCTestCase {
 
         // Given
         let request: ImmutableRequest = .mockWith(httpMethod: "POST")
-        let interception = URLSessionTaskInterception(request: request, isFirstParty: true)
+        // TODO: RUM-13455 - Trace requires metrics mode for now
+        let interception = URLSessionTaskInterception(request: request, isFirstParty: true, trackingMode: .metrics)
         interception.register(response: .mockResponseWith(statusCode: 200), error: nil)
         interception.register(
             metrics: .mockWith(
@@ -326,7 +329,8 @@ class TracingURLSessionHandlerTests: XCTestCase {
         core.onEventWriteContext = { _ in expectation.fulfill() }
 
         // Given
-        let incompleteInterception = URLSessionTaskInterception(request: .mockAny(), isFirstParty: true)
+        // TODO: RUM-13455 - Trace requires metrics mode for now
+        let incompleteInterception = URLSessionTaskInterception(request: .mockAny(), isFirstParty: true, trackingMode: .metrics)
         // `incompleteInterception` has no metrics and no completion
 
         // When
@@ -343,7 +347,8 @@ class TracingURLSessionHandlerTests: XCTestCase {
         core.onEventWriteContext = { _ in expectation.fulfill() }
 
         // Given
-        let interception = URLSessionTaskInterception(request: .mockAny(), isFirstParty: false)
+        // TODO: RUM-13455 - Trace requires metrics mode for now
+        let interception = URLSessionTaskInterception(request: .mockAny(), isFirstParty: false, trackingMode: .metrics)
         interception.register(response: .mockAny(), error: nil)
         interception.register(
             metrics: .mockWith(
@@ -371,7 +376,8 @@ class TracingURLSessionHandlerTests: XCTestCase {
         let request: ImmutableRequest = .mockWith(
             allHTTPHeaderFields: [TracingHTTPHeaders.originField: "rum"]
         )
-        let interception = URLSessionTaskInterception(request: request, isFirstParty: false)
+        // TODO: RUM-13455 - Trace requires metrics mode for now
+        let interception = URLSessionTaskInterception(request: request, isFirstParty: false, trackingMode: .metrics)
         interception.register(response: .mockAny(), error: nil)
 
         // When
@@ -387,7 +393,8 @@ class TracingURLSessionHandlerTests: XCTestCase {
         core.onEventWriteContext = { _ in expectation.fulfill() }
 
         // Given
-        let interception = URLSessionTaskInterception(request: .mockAny(), isFirstParty: true)
+        // TODO: RUM-13455 - Trace requires metrics mode for now
+        let interception = URLSessionTaskInterception(request: .mockAny(), isFirstParty: true, trackingMode: .metrics)
         interception.register(response: .mockAny(), error: nil)
         interception.register(
             metrics: .mockWith(
@@ -427,7 +434,8 @@ class TracingURLSessionHandlerTests: XCTestCase {
 
         core.context.applicationStateHistory = .mockAppInForeground()
 
-        let interception = URLSessionTaskInterception(request: .mockAny(), isFirstParty: true)
+        // TODO: RUM-13455 - Trace requires metrics mode for now
+        let interception = URLSessionTaskInterception(request: .mockAny(), isFirstParty: true, trackingMode: .metrics)
         interception.register(response: .mockAny(), error: nil)
         interception.register(
             metrics: .mockWith(